+++ doc/diary/2016_02_28.md
@@ -114,9 +114,9 @@ After execution the outcome needs to be analysed in the **execution network**.
 E.g. when an error has occured, the given strategy needs to be downvoted. A
 more complex task is to upvote the score of a strategy in case it succeeds,
 because the information if a strategy succeeds relies in the first place on the
-reaction of the putput we are going to send out. Here the decisions rely on
+reaction of the output we are going to send out. Here the decisions rely on
 information based on inputs across requests. So the evaluation network needs to
 track recent strategies and their effective outcome to evaluate success or
-failure of the currently chosen one.
+failure of strategies.
 
 xh3b4sd
