+++ factory/factory.go
@@ -1,2 +1,3 @@
-// TODO
+// Package factory implements spec.Factory to create new objects on demand. The
+// factory concept solves the problem of dependency injection.
 package factory
