+++ file-system/file_system.go
@@ -1,2 +1,3 @@
-// Package filesystem implements spec.FileSystem.
+// Package filesystem implements spec.FileSystem to provide a filesystem
+// abstraction especially for testing reasons.
 package filesystem
