+++ spec/network.go
@@ -1,5 +1,11 @@
 package spec
 
+// Network implementations provide some sort of higher level business logic.
+// They are also used to group and manage neurons. Networks here should not be
+// too much aligned with the biological equivalent. Recent implementations
+// turned out to be way too complex and stood in the way of getting things
+// done. Anyway, networks need to provide critical business logic though. At
+// the end the question is what makes sense and what works out.
 type Network interface {
 	Object
