+++ spec/neuron.go
@@ -1,5 +1,12 @@
 package spec
 
+// Neuron implementations provide some sort of specialized business logic. They
+// are used to make networks dynamic and functional. Neurons here should not be
+// too much aligned with the biological equivalent. Recent implementations
+// turned out to be way too complex and stood in the way of getting things
+// done. Anyway, neurons can be used for very specific very tiny tasks though.
+// At the end the question is what makes sense and what works out. Neurons
+// simply represent a pragmatic way to separate concerns where appropriate.
 type Neuron interface {
 	Object
